Extract API call logic into a weather.py file
